--- src/main/java/core/HttpResponse.java
@@ -1,19 +1,0 @@
-package core;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.net.Socket;
-
-public class HttpResponse {
-
-    private OutputStream out;
-
-    public HttpResponse(OutputStream out) {
-        this.out = out;
-    }
-
-    public OutputStream getOut() {
-        return out;
-    }
-
-}
