--- src/main/java/handler/mapping/GetMapping.java
@@ -1,14 +1,0 @@
-package handler.mapping;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface GetMapping {
-
-    String url() default "";
-
-}
