+++ src/main/java/util/HttpRequestUtils.java
@@ -6,22 +6,8 @@ import java.util.stream.Collectors;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.Maps;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class HttpRequestUtils {
-
-    private static final Logger log = LoggerFactory.getLogger(HttpRequestUtils.class);
-
-    public static String getUrl(String firstLine) {
-        String[] tokens = firstLine.split(" ");
-        if(tokens.length != 3) {
-            //TODO : RETURN HTTP STATUS 500 ?
-        }
-        return tokens[1];
-    }
-
-
     /**
      * @param queryString은
      *            URL에서 ? 이후에 전달되는 field1=value1&field2=value2 형식임
