--- src/main/java/util/ResponseData.java
@@ -1,67 +1,0 @@
-package util;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.DataOutputStream;
-import java.io.IOException;
-
-public class ResponseData {
-
-    private static final Logger log = LoggerFactory.getLogger(ResponseData.class);
-
-    public static void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-
-    public static void response200HeaderWithCss(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/css;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-
-    public static void response302Header(DataOutputStream dos, String location) {
-        log.debug("location : {}", location);
-        try {
-            dos.writeBytes("HTTP/1.1 302 Found \r\n"); // 원래는 200
-            dos.writeBytes("Location: " + location+"\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-
-    public static void response302HeaderWithCookie(DataOutputStream dos, String location, String cookie) {
-        log.debug("location : {}", location);
-        try {
-            dos.writeBytes("HTTP/1.1 302 Found \r\n"); // 원래는 200
-            dos.writeBytes("Location: " + location+"\r\n");
-            dos.writeBytes("Set-Cookie: " + cookie + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-
-    public static void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-    
-}
