+++ src/main/java/webserver/WebServer.java
@@ -26,8 +26,8 @@ public class WebServer {
             // 클라이언트가 연결될때까지 대기한다.
             Socket connection;
             while ((connection = listenSocket.accept()) != null) {
-                WebCat webCat = new WebCat(connection);
-                webCat.start();
+                RequestHandler requestHandler = new RequestHandler(connection);
+                requestHandler.start();
             }
         }
     }
