+++ src/test/java/util/IOUtilsTest.java
@@ -3,7 +3,7 @@ package util;
 import java.io.BufferedReader;
 import java.io.StringReader;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
